Evaluation/codes/requirements.txt
# Core dependencies
torch>=2.9.1
transformers>=4.57.1
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Qwen2-VL specific
qwen-vl-utils

# Evaluation metrics
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.3.1
bert-score>=0.3.13

# Visualizations
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional for better performance
accelerate>=0.20.0
einops>=0.7.0
timm>=0.9.0
